+++ packages/react/src/index.ts
@@ -2,6 +2,7 @@
 
 export { createReactSpoosh, type SpooshReactHooks } from "./createReactSpoosh";
 export { createUseRead } from "./useRead";
+export { createUseLazyRead } from "./useLazyRead";
 export { createUseWrite } from "./useWrite";
 export { createUseInfiniteRead } from "./useInfiniteRead";
 
@@ -12,6 +13,7 @@ export type {
   UseWriteResult,
   UseInfiniteReadResult,
   BaseReadResult,
+  BaseLazyReadResult,
   BaseWriteResult,
   BaseInfiniteReadResult,
   BaseInfiniteReadOptions,